// packages/react-native-dogfood/gen/sfu_events/events.ts
export interface SfuEvent {
    /**
     * @generated from protobuf oneof: event_payload
     */
    eventPayload: {
        oneofKind: "subscriberOffer";
        /**
         * @generated from protobuf field: stream.video.sfu.SubscriberOffer subscriber_offer = 1;
         */
        subscriberOffer: SubscriberOffer;
    } | {
        oneofKind: "connectionQualityChanged";
        /**
         * @generated from protobuf field: stream.video.sfu.ConnectionQualityChanged connection_quality_changed = 2;
         */
        connectionQualityChanged: ConnectionQualityChanged;
    } | {
        oneofKind: "audioLevelChanged";
        /**
         * @generated from protobuf field: stream.video.sfu.AudioLevelChanged audio_level_changed = 3;
         */
        audioLevelChanged: AudioLevelChanged;
    } | {
        oneofKind: "subscriberCandidate";
        /**
         * @generated from protobuf field: stream.video.sfu.ICECandidateTrickle subscriber_candidate = 4;
         */
        subscriberCandidate: ICECandidateTrickle;
    } | {
        oneofKind: "publisherCandidate";
        /**
         * @generated from protobuf field: stream.video.sfu.ICECandidateTrickle publisher_candidate = 5;
         */
        publisherCandidate: ICECandidateTrickle;
    } | {
        oneofKind: "changePublishQuality";
        /**
         * @generated from protobuf field: stream.video.sfu.ChangePublishQuality change_publish_quality = 6;
         */
        changePublishQuality: ChangePublishQuality;
    } | {
        oneofKind: "localDeviceChange";
        /**
         * @generated from protobuf field: stream.video.sfu.LocalDeviceChange local_device_change = 7;
         */
        localDeviceChange: LocalDeviceChange;
    } | {
        oneofKind: "muteStateChanged";
        /**
         * @generated from protobuf field: stream.video.sfu.MuteStateChanged mute_state_changed = 8;
         */
        muteStateChanged: MuteStateChanged;
    } | {
        oneofKind: "videoQualityChanged";
        /**
         * @generated from protobuf field: stream.video.sfu.VideoQualityChanged video_quality_changed = 9;
         */
        videoQualityChanged: VideoQualityChanged;
    } | {
        oneofKind: "participantJoined";
        /**
         * @generated from protobuf field: stream.video.sfu.ParticipantJoined participant_joined = 10;
         */
        participantJoined: ParticipantJoined;
    } | {
        oneofKind: "participantLeft";
        /**
         * @generated from protobuf field: stream.video.sfu.ParticipantLeft participant_left = 11;
         */
        participantLeft: ParticipantLeft;
    } | {
        oneofKind: "dominantSpeakerChanged";
        /**
         * @generated from protobuf field: stream.video.sfu.DominantSpeakerChanged dominant_speaker_changed = 12;
         */
        dominantSpeakerChanged: DominantSpeakerChanged;
    } | {
        oneofKind: undefined;
    };
}
/**
 * ParticipantJoined is fired when a user joins a call
 *
 * @generated from protobuf message stream.video.sfu.ParticipantJoined
 */
export interface ParticipantJoined {
    /**
     * @generated from protobuf field: stream.video.sfu.Call call = 1;
     */
    call?: Call;
    /**
     * @generated from protobuf field: stream.video.sfu.Participant participant = 2;
     */
    participant?: Participant;
}
/**
 * ParticipantJoined is fired when a user leaves a call
 *
 * @generated from protobuf message stream.video.sfu.ParticipantLeft
 */
export interface ParticipantLeft {
    /**
     * @generated from protobuf field: stream.video.sfu.Call call = 1;
     */
    call?: Call;
    /**
     * @generated from protobuf field: stream.video.sfu.Participant participant = 2;
     */
    participant?: Participant;
}
/**
 * @generated from protobuf message stream.video.sfu.MuteStateChanged
 */
export interface MuteStateChanged {
    /**
     * @generated from protobuf field: string user_id = 1;
     */
    userId: string;
    /**
     * @generated from protobuf field: bool audio_muted = 2;
     */
    audioMuted: boolean;
    /**
     * @generated from protobuf field: bool video_muted = 3;
     */
    videoMuted: boolean;
}
/**
 * TODO batch these messages
 *
 * @generated from protobuf message stream.video.sfu.VideoQualityChanged
 */
export interface VideoQualityChanged {
    /**
     * @generated from protobuf field: repeated stream.video.sfu.StreamQuality stream_qualities = 1;
     */
    streamQualities: StreamQuality[];
}
/**
 * SubscriberOffer is sent when the SFU adds tracks to a subscription
 * this usually happens when a user joins a call and starts sending video
 *
 * @generated from protobuf message stream.video.sfu.SubscriberOffer
 */
export interface SubscriberOffer {
    /**
     * @generated from protobuf field: string sdp = 1;
     */
    sdp: string;
}
/**
 * @generated from protobuf message stream.video.sfu.LocalDeviceChange
 */
export interface LocalDeviceChange {
    /**
     * @generated from protobuf field: string type = 1;
     */
    type: string;
}
/**
 * ICECandidateTrickle is about the ICE nonsense stuff
 *
 * @generated from protobuf message stream.video.sfu.ICECandidateTrickle
 */
export interface ICECandidateTrickle {
    /**
     * @generated from protobuf field: string candidate = 1;
     */
    candidate: string;
}
/**
 * ConnectionQuality is sent to inform about connection quality changes
 * eg. thierry's connection is not good -> render a red icon Zoom style
 *
 * @generated from protobuf message stream.video.sfu.ConnectionQualityChanged
 */
export interface ConnectionQualityChanged {
    /**
     * @generated from protobuf field: string user_id = 1;
     */
    userId: string;
    /**
     * @generated from protobuf field: stream.video.sfu.ConnectionQuality connection_quality = 2;
     */
    connectionQuality: ConnectionQuality;
}
/**
 * DominantSpeakerChanged is sent by the SFU to notify when there is a new dominant speaker in the call
 *
 * @generated from protobuf message stream.video.sfu.DominantSpeakerChanged
 */
export interface DominantSpeakerChanged {
    /**
     * @generated from protobuf field: string user_id = 1;
     */
    userId: string;
}
/**
 * AudioLevelChanged is sent by the SFU to notify about audio levels by user
 *
 * @generated from protobuf message stream.video.sfu.AudioLevelChanged
 */
export interface AudioLevelChanged {
    /**
     * @generated from protobuf field: repeated stream.video.sfu.AudioLevel audio_levels = 1;
     */
    audioLevels: AudioLevel[];
}
/**
 * @generated from protobuf message stream.video.sfu.AudioLevel
 */
export interface AudioLevel {
    /**
     * @generated from protobuf field: string user_id = 1;
     */
    userId: string;
    /**
     * 0.0 means complete silence, 1.0 loudest
     *
     * @generated from protobuf field: float level = 2;
     */
    level: number;
}
// ...
export interface AudioSender {
    /**
     * @generated from protobuf field: stream.video.sfu.AudioMediaRequest mediaRequest = 1;
     */
    mediaRequest?: AudioMediaRequest;
    /**
     * @generated from protobuf field: stream.video.sfu.Codec codec = 2;
     */
    codec?: Codec;
    /**
     * @generated from protobuf field: repeated stream.video.sfu.AudioLayerSetting layers = 3;
     */
    layers: AudioLayerSetting[];
}
// ...
export interface VideoSender {
    /**
     * @generated from protobuf field: stream.video.sfu.VideoMediaRequest mediaRequest = 1;
     */
    mediaRequest?: VideoMediaRequest;
    /**
     * @generated from protobuf field: stream.video.sfu.Codec codec = 2;
     */
    codec?: Codec;
    /**
     * @generated from protobuf field: repeated stream.video.sfu.VideoLayerSetting layers = 3;
     */
    layers: VideoLayerSetting[];
}
/**
 * sent to users when they need to change the quality of their video
 *
 * @generated from protobuf message stream.video.sfu.ChangePublishQuality
 */
export interface ChangePublishQuality {
    /**
     * @generated from protobuf field: repeated stream.video.sfu.AudioSender audio_sender = 1;
     */
    audioSender: AudioSender[];
    /**
     * @generated from protobuf field: repeated stream.video.sfu.VideoSender video_sender = 2;
     */
    videoSender: VideoSender[];
}
// ...
export const SfuEvent = new SfuEvent$Type();
// ...
export const ParticipantJoined = new ParticipantJoined$Type();
// ...
export const ParticipantLeft = new ParticipantLeft$Type();
// ...
export const MuteStateChanged = new MuteStateChanged$Type();
// ...
export const VideoQualityChanged = new VideoQualityChanged$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SubscriberOffer$Type extends MessageType<SubscriberOffer> {
    constructor() {
        super("stream.video.sfu.SubscriberOffer", [
            { no: 1, name: "sdp", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<SubscriberOffer>): SubscriberOffer {
        const message = { sdp: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<SubscriberOffer>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SubscriberOffer): SubscriberOffer {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string sdp */ 1:
                    message.sdp = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SubscriberOffer, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string sdp = 1; */
        if (message.sdp !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sdp);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message stream.video.sfu.SubscriberOffer
 */
export const SubscriberOffer = new SubscriberOffer$Type();
// ...
export const LocalDeviceChange = new LocalDeviceChange$Type();
// ...
export const ICECandidateTrickle = new ICECandidateTrickle$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ConnectionQualityChanged$Type extends MessageType<ConnectionQualityChanged> {
    constructor() {
        super("stream.video.sfu.ConnectionQualityChanged", [
            { no: 1, name: "user_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "connection_quality", kind: "enum", T: () => ["stream.video.sfu.ConnectionQuality", ConnectionQuality] }
        ]);
    }
    create(value?: PartialMessage<ConnectionQualityChanged>): ConnectionQualityChanged {
        const message = { userId: "", connectionQuality: 0 };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<ConnectionQualityChanged>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ConnectionQualityChanged): ConnectionQualityChanged {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string user_id */ 1:
                    message.userId = reader.string();
                    break;
                case /* stream.video.sfu.ConnectionQuality connection_quality */ 2:
                    message.connectionQuality = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ConnectionQualityChanged, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string user_id = 1; */
        if (message.userId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.userId);
        /* stream.video.sfu.ConnectionQuality connection_quality = 2; */
        if (message.connectionQuality !== 0)
            writer.tag(2, WireType.Varint).int32(message.connectionQuality);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message stream.video.sfu.ConnectionQualityChanged
 */
export const ConnectionQualityChanged = new ConnectionQualityChanged$Type();
// ...
export const DominantSpeakerChanged = new DominantSpeakerChanged$Type();
// ...
export const AudioLevelChanged = new AudioLevelChanged$Type();
// ...
export const AudioLevel = new AudioLevel$Type();
// ...
export const AudioSender = new AudioSender$Type();
// ...
export const VideoSender = new VideoSender$Type();
// ...
export const ChangePublishQuality = new ChangePublishQuality$Type();